pyproject.toml
[project]
name = "arpias-square-helper"
version = "0.1.0"
description = "CLI tool to capture Square session cookies and upload to Arpias server"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pydoll-python>=2.14.0",
    "httpx>=0.25.0",
    "click>=8.1.0",
]

[project.scripts]
arpias-square-helper = "square_login.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["square_login"]
